According to google groups, the toString should not return a typeCased numerator that is a double. Changes reflect that

+++ RationalNumber.java
@@ -61,7 +61,7 @@ public class RationalNumber extends RealNumber
   *@return the value expressed as "3/4" or "8/3"
   */
   public String toString(){
-    if (denominator == 1 || numerator == 0) return "" + (double) numerator + "";
+    if (denominator == 1 || numerator == 0) return "" + numerator + "";
     else return numerator + "/" + denominator;
 }
